Remove Guice, upgrade library versions: log4j, slf4j, reflections, mockito.
