+++ pom.xml
@@ -32,16 +32,13 @@
     <properties>
         <jdk.version>1.8</jdk.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <!--lucene.version>${project.version}</lucene.version-->
         <lucene.version>8.0.0</lucene.version>
         <hadoop.version>2.7.2</hadoop.version>
-        <log4j.version>2.11.0</log4j.version>
-        <slf4j.version>1.7.7</slf4j.version>
-        <guice.version>4.1.0</guice.version>
-        <guava.version>14.0.1</guava.version>
-        <reflections.version>0.9.5</reflections.version>
+        <log4j.version>2.11.2</log4j.version>
+        <slf4j.version>1.7.24</slf4j.version>
+        <reflections.version>0.9.11</reflections.version>
         <junit.version>4.12</junit.version>
-        <mockito.version>2.6.2</mockito.version>
+        <mockito.version>2.23.4</mockito.version>
         <findbugs.version>3.0.1</findbugs.version>
         <jgoodies.version>2.7.0</jgoodies.version>
         <firbiddenapi.version>2.6</firbiddenapi.version>
@@ -201,16 +198,6 @@
             <groupId>com.google.code.findbugs</groupId>
             <artifactId>jsr305</artifactId>
             <version>${findbugs.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.google.inject</groupId>
-            <artifactId>guice</artifactId>
-            <version>${guice.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <version>${guava.version}</version>
         </dependency>
         <dependency>
             <groupId>org.reflections</groupId>
