+++ src/main/java/org/apache/lucene/luke/app/desktop/components/LogsPanelProvider.java
@@ -17,9 +17,6 @@
 
 package org.apache.lucene.luke.app.desktop.components;
 
-import com.google.inject.Inject;
-import com.google.inject.Provider;
-import com.google.inject.name.Named;
 import org.apache.lucene.luke.app.desktop.util.MessageUtils;
 
 import javax.swing.BorderFactory;
@@ -31,18 +28,16 @@ import java.awt.BorderLayout;
 import java.awt.FlowLayout;
 import java.io.File;
 
-public final class LogsPanelProvider implements Provider<JPanel> {
+public final class LogsPanelProvider {
 
   private static final String LOG_FILE = System.getProperty("user.home") + File.separator + ".luke.d" + File.separator + "luke.log";
 
   private final JTextArea logTextArea;
 
-  @Inject
-  public LogsPanelProvider(@Named("log_area") JTextArea logTextArea) {
+  public LogsPanelProvider(JTextArea logTextArea) {
     this.logTextArea = logTextArea;
   }
 
-  @Override
   public JPanel get() {
     JPanel panel = new JPanel(new BorderLayout());
     panel.setOpaque(false);
