+++ src/main/java/org/apache/lucene/luke/app/desktop/components/dialog/ConfirmDialogFactory.java
@@ -17,11 +17,102 @@
 
 package org.apache.lucene.luke.app.desktop.components.dialog;
 
+import org.apache.lucene.luke.app.desktop.Preferences;
+import org.apache.lucene.luke.app.desktop.PreferencesFactory;
 import org.apache.lucene.luke.app.desktop.util.DialogOpener;
+import org.apache.lucene.luke.app.desktop.util.FontUtils;
+import org.apache.lucene.luke.app.desktop.util.MessageUtils;
 import org.apache.lucene.luke.app.desktop.util.lang.Callable;
 
-public interface ConfirmDialogFactory extends DialogOpener.DialogFactory {
-  void setMessage(String message);
+import javax.swing.BorderFactory;
+import javax.swing.JButton;
+import javax.swing.JDialog;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+import java.awt.BorderLayout;
+import java.awt.Color;
+import java.awt.Dialog;
+import java.awt.Dimension;
+import java.awt.FlowLayout;
+import java.awt.Font;
+import java.awt.GridLayout;
+import java.awt.Window;
+import java.io.IOException;
 
-  void setCallback(Callable callback);
+public final class ConfirmDialogFactory implements DialogOpener.DialogFactory {
+
+  private static ConfirmDialogFactory instance;
+
+  private final Preferences prefs;
+
+  private JDialog dialog;
+
+  private String message;
+
+  private Callable callback;
+
+  public synchronized static ConfirmDialogFactory getInstance() throws IOException {
+    if (instance == null) {
+      instance = new ConfirmDialogFactory();
+    }
+    return instance;
+  }
+
+  private ConfirmDialogFactory() throws IOException {
+    this.prefs = PreferencesFactory.getInstance();
+  }
+
+  public void setMessage(String message) {
+    this.message = message;
+  }
+
+  public void setCallback(Callable callback) {
+    this.callback = callback;
+  }
+
+  @Override
+  public JDialog create(Window owner, String title, int width, int height) {
+    dialog = new JDialog(owner, title, Dialog.ModalityType.APPLICATION_MODAL);
+    dialog.add(content());
+    dialog.setSize(new Dimension(width, height));
+    dialog.setLocationRelativeTo(owner);
+    dialog.getContentPane().setBackground(prefs.getColorTheme().getBackgroundColor());
+    return dialog;
+  }
+
+  private JPanel content() {
+    JPanel panel = new JPanel(new BorderLayout());
+    panel.setOpaque(false);
+    panel.setBorder(BorderFactory.createEmptyBorder(15, 15, 15, 15));
+
+    JPanel header = new JPanel(new FlowLayout(FlowLayout.LEADING));
+    header.setOpaque(false);
+    JLabel alertIconLbl = new JLabel(FontUtils.elegantIconHtml("&#x71;"));
+    alertIconLbl.setHorizontalAlignment(JLabel.CENTER);
+    alertIconLbl.setFont(new Font(alertIconLbl.getFont().getFontName(), Font.PLAIN, 25));
+    header.add(alertIconLbl);
+    panel.add(header, BorderLayout.PAGE_START);
+
+    JPanel center = new JPanel(new GridLayout(1, 1));
+    center.setOpaque(false);
+    center.setBorder(BorderFactory.createLineBorder(Color.gray, 3));
+    center.add(new JLabel(message, JLabel.CENTER));
+    panel.add(center, BorderLayout.CENTER);
+
+    JPanel footer = new JPanel(new FlowLayout(FlowLayout.TRAILING));
+    footer.setOpaque(false);
+    JButton okBtn = new JButton(MessageUtils.getLocalizedMessage("button.ok"));
+    okBtn.addActionListener(e -> {
+      callback.call();
+      dialog.dispose();
+    });
+    footer.add(okBtn);
+    JButton closeBtn = new JButton(MessageUtils.getLocalizedMessage("button.close"));
+    closeBtn.addActionListener(e -> dialog.dispose());
+    footer.add(closeBtn);
+    panel.add(footer, BorderLayout.PAGE_END);
+
+    return panel;
+  }
+
 }
