+++ src/main/java/org/apache/lucene/luke/models/tools/IndexToolsImpl.java
@@ -35,8 +35,8 @@ import org.apache.lucene.index.CheckIndex;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
-import org.apache.lucene.luke.models.LukeModel;
 import org.apache.lucene.luke.models.LukeException;
+import org.apache.lucene.luke.models.LukeModel;
 import org.apache.lucene.luke.util.IndexUtils;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.store.Directory;
